MB-19667 - consult base segment during merge ops
The merger goroutine would use an iterator when merging or squashing
multiple segments together.

When that iterator encounters MergeOperators, it would just do a Get()
to resolve all the MergeOperators into a single value.

But, that Get() would incorrectly not consult the stackDirtyBase, so
any MergeOperators in the stackDirtyBase would be incorrectly missed.

Change-Id: I808878af674cfcd63025f3e082c771c65cddd05a